Move sign-in/pinning code out of SceneEntryManager
